don't use complex.h
C complex types don't mix well with C++ and GCC >=6

The source of the problem is the "overlay" version of many of the C standard headers
(<complex.h> among them) which GCC installs.  eg. "/usr/include/c++/8/complex.h"
which is earlier in the implied search path than "/usr/include/complex.h".

This has long been the case, and can work through the gcc specific "#include_next"
directive to also include the real header.  The recent (circa gcc 6) change is to skip
the "#include_next" when compiling c++.  This hides the C99 standard complex types,
and breaks the magic which FFTW uses to decide whether to use the C99 complex type,
or the older 'double[2]' array type.